payment method fixing

+++ Lab-11/KenaKata/src/main/java/kenakata/payment/PaymentMethod.java
@@ -1,5 +1,5 @@
 package kenakata.payment;
 
 public interface PaymentMethod {
-     void authorise(int amount);
+     void authorise(long amount);
 }
